Relocate ETVM's MoveToIndex/SortByName tree-sort helpers into Gum.Presentation, part of #3845
They already operated purely through ITreeNodeMutable/ITreeNode (no WinForms types), a byproduct
of #3841's reparent/reorder conversion - so this is a pure file-location move, not a behavior
change. Pinned by the existing TreeNodeMutationExtensionsTests (GumToolUnitTests), unchanged and
green before/after.
