+++ Gum/Plugins/InternalPlugins/TreeView/ElementTreeViewManager.cs
@@ -2891,119 +2891,9 @@ public static class TreeNodeExtensionMethods
             treeNode.Parent.IsTopComponentContainerTreeNode());
     }
 
-    /// <summary>
-    /// Moves <paramref name="node"/> to <paramref name="desiredIndex"/> among its current parent's
-    /// children. No-op if it is already there, or if it currently has no parent.
-    /// </summary>
-    /// <remarks>
-    /// Shared by ElementTreeViewManager's element/instance and behavior-instance reorder paths
-    /// (<c>RefreshElementTreeNode</c>/<c>RefreshBehaviorTreeNode</c>), which both reorder a node
-    /// within its already-correct parent after any reparenting has already happened.
-    /// </remarks>
-    public static void MoveToIndex(this ITreeNodeMutable node, int desiredIndex)
-    {
-        ITreeNodeMutable? parent = node.Parent;
-        if (parent == null || parent.IndexOfChild(node) == desiredIndex)
-        {
-            return;
-        }
-
-        parent.RemoveChild(node);
-        parent.InsertChild(desiredIndex, node);
-    }
-
-    /// <summary>
-    /// Sorts a node's direct children alphabetically by name, with folders appearing before files.
-    /// </summary>
-    /// <param name="parentNode">The node whose children should be sorted.</param>
-    /// <param name="recursive">
-    /// If true, recursively sorts all child node collections (except within Screen, Component, Standard, or Behavior element nodes).
-    /// Default is false.
-    /// </param>
-    /// <remarks>
-    /// The sort order places folders (Components and Screens subfolders) before individual elements,
-    /// and within each category, nodes are sorted alphabetically by their Text property.
-    /// When recursive is true, the method will not sort children of element nodes (Screen, Component, Standard, Behavior)
-    /// as these typically contain instances and states that should maintain their specific order.
-    /// </remarks>
-    public static void SortByName(this ITreeNodeMutable parentNode, bool recursive = false)
-    {
-        int lastObjectExclusive = parentNode.ChildCount;
-        int whereObjectBelongs;
-        for (int i = 0 + 1; i < lastObjectExclusive; i++)
-        {
-            ITreeNodeMutable first = parentNode.GetChildAt(i);
-            ITreeNodeMutable second = parentNode.GetChildAt(i - 1);
-            if (FirstComesBeforeSecond(first, second))
-            {
-                if (i == 1)
-                {
-                    ITreeNodeMutable movingNode = parentNode.GetChildAt(i);
-                    parentNode.RemoveChildAt(i);
-
-                    parentNode.InsertChild(0, movingNode);
-                    continue;
-                }
-
-                for (whereObjectBelongs = i - 2; whereObjectBelongs > -1; whereObjectBelongs--)
-                {
-                    second = parentNode.GetChildAt(whereObjectBelongs);
-                    if (!FirstComesBeforeSecond(parentNode.GetChildAt(i), second))
-                    {
-                        ITreeNodeMutable movingNode = parentNode.GetChildAt(i);
-
-                        parentNode.RemoveChildAt(i);
-                        parentNode.InsertChild(whereObjectBelongs + 1, movingNode);
-                        break;
-                    }
-                    else if (whereObjectBelongs == 0 && FirstComesBeforeSecond(parentNode.GetChildAt(i), parentNode.GetChildAt(0)))
-                    {
-                        ITreeNodeMutable movingNode = parentNode.GetChildAt(i);
-                        parentNode.RemoveChildAt(i);
-                        parentNode.InsertChild(0, movingNode);
-                        break;
-                    }
-                }
-            }
-        }
-
-        if(recursive)
-        {
-            for (int i = 0; i < parentNode.ChildCount; i++)
-            {
-                ITreeNodeMutable childNode = parentNode.GetChildAt(i);
-
-                var sortInner = childNode.IsScreenTreeNode() == false &&
-                    childNode.IsComponentTreeNode() == false &&
-                    childNode.IsStandardElementTreeNode() == false &&
-                    childNode.IsBehaviorTreeNode() == false;
-
-                if(sortInner)
-                {
-                    childNode.SortByName(recursive);
-                }
-            }
-        }
-    }
-
-    private static bool FirstComesBeforeSecond(ITreeNodeMutable first, ITreeNodeMutable second)
-    {
-        bool isFirstDirectory = first.IsComponentsFolderTreeNode() || first.IsScreensFolderTreeNode();
-        bool isSecondDirectory = second.IsComponentsFolderTreeNode() || second.IsScreensFolderTreeNode();
-
-        if (isFirstDirectory && !isSecondDirectory)
-        {
-            return true;
-        }
-        else if (!isFirstDirectory && isSecondDirectory)
-        {
-            return false;
-        }
-        else
-        {
-            return first.Text.CompareTo(second.Text) < 0;
-        }
-    }
+    // MoveToIndex/SortByName/FirstComesBeforeSecond moved to Gum.Presentation's
+    // TreeNodeMutationExtensions (part of #3845) - they only touched ITreeNodeMutable/ITreeNode
+    // members already, so this was a pure file-location move. See that file for the implementation.
 
     /// <summary>
     /// Gets all descendant nodes of the tree node in a flattened list, recursively traversing the entire tree structure.
